requirements.txt
anyio==3.6.2
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
flake8==4.0.1
h11==0.14.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==4.2.0
iniconfig==1.1.1
itsdangerous==2.1.1
lxml==4.9.2
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.1.0
python-telegram-bot==20.0
requests==2.28.2
requests-cache==0.6.3
requests-mock==1.9.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.61.0
typing_extensions==4.1.1
url-normalize==1.4.3
urllib3==1.26.14
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.7.0
pyaes==1.6.1
pyrogram==2.0.97
pysocks==1.7.1
